Extract EOBaseProp
