+++ transforms/helpers/eo-prop/index.ts
@@ -1,10 +1,12 @@
 import type { Property } from 'jscodeshift';
 import type { ImportPropDecoratorMap } from '../decorator-info';
 import type { RuntimeData } from '../runtime-data';
+import { assert } from '../util/types';
+import type { default as EOProp } from './private/abstract';
 import EOActionsObjectProp, {
   isEOActionsPropProperty,
 } from './private/actions-object';
-import EOProp from './private/base';
+import EOBaseProp, { isEOProperty } from './private/base';
 import EOCallExpressionProp, {
   isCallExpressionProperty,
 } from './private/call-expression';
@@ -15,8 +17,9 @@ import EOFunctionExpressionProp, {
   isFunctionExpressionProperty,
 } from './private/function-expression';
 
-export { default as EOProp } from './private/base';
+export type { default as EOProp } from './private/abstract';
 export { default as EOActionsObjectProp } from './private/actions-object';
+export { default as EOBaseProp } from './private/base';
 export { default as EOCallExpressionProp } from './private/call-expression';
 export { default as EOClassDecoratorProp } from './private/class-decorator';
 export { default as EOFunctionExpressionProp } from './private/function-expression';
@@ -44,6 +47,7 @@ export default function makeEOProp(
   } else if (isEOActionsPropProperty(eoProp)) {
     return new EOActionsObjectProp(eoProp, runtimeData);
   } else {
-    return new EOProp(eoProp, runtimeData);
+    assert(isEOProperty(eoProp));
+    return new EOBaseProp(eoProp, runtimeData);
   }
 }
